Moved LevelComparison to Core namespace
