+++ Forge/Effects/Stacking/StackingData.cs
@@ -1,5 +1,6 @@
 // Copyright © Gamesmiths Guild.
 
+using Gamesmiths.Forge.Core;
 using Gamesmiths.Forge.Effects.Magnitudes;
 
 namespace Gamesmiths.Forge.Effects.Stacking;
